Finish separating out `ReplStore`
