+++ crates/repl/src/repl_store.rs
@@ -1,28 +1,16 @@
-use std::ops::Range;
 use std::sync::Arc;
 
-use anyhow::{Context as _, Result};
+use anyhow::Result;
 use collections::HashMap;
-use editor::{Anchor, Editor, RangeToAnchorExt};
 use gpui::{
     prelude::*, AppContext, EntityId, Global, Model, ModelContext, Subscription, Task, View,
-    ViewContext, WeakView,
 };
-use language::{Language, Point};
-use multi_buffer::MultiBufferRow;
+use language::Language;
 use project::Fs;
 use settings::{Settings, SettingsStore};
 
 use crate::kernels::kernel_specifications;
-use crate::session::SessionEvent;
 use crate::{JupyterSettings, KernelSpecification, Session};
-
-pub enum SessionSupport {
-    ActiveSession(View<Session>),
-    Inactive(Box<KernelSpecification>),
-    RequiresSetup(Arc<str>),
-    Unsupported,
-}
 
 struct GlobalReplStore(Model<ReplStore>);
 
@@ -80,63 +68,6 @@ impl ReplStore {
         }
     }
 
-    pub fn snippet(
-        &self,
-        editor: WeakView<Editor>,
-        cx: &mut ViewContext<Self>,
-    ) -> Option<(String, Arc<Language>, Range<Anchor>)> {
-        let editor = editor.upgrade()?;
-        let editor = editor.read(cx);
-
-        let buffer = editor.buffer().read(cx).snapshot(cx);
-
-        let selection = editor.selections.newest::<usize>(cx);
-        let multi_buffer_snapshot = editor.buffer().read(cx).snapshot(cx);
-
-        let range = if selection.is_empty() {
-            let cursor = selection.head();
-
-            let cursor_row = multi_buffer_snapshot.offset_to_point(cursor).row;
-            let start_offset = multi_buffer_snapshot.point_to_offset(Point::new(cursor_row, 0));
-
-            let end_point = Point::new(
-                cursor_row,
-                multi_buffer_snapshot.line_len(MultiBufferRow(cursor_row)),
-            );
-            let end_offset = start_offset.saturating_add(end_point.column as usize);
-
-            // Create a range from the start to the end of the line
-            start_offset..end_offset
-        } else {
-            selection.range()
-        };
-
-        let anchor_range = range.to_anchors(&multi_buffer_snapshot);
-
-        let selected_text = buffer
-            .text_for_range(anchor_range.clone())
-            .collect::<String>();
-
-        let start_language = buffer.language_at(anchor_range.start)?;
-        let end_language = buffer.language_at(anchor_range.end)?;
-        if start_language != end_language {
-            return None;
-        }
-
-        Some((selected_text, start_language.clone(), anchor_range))
-    }
-
-    pub fn language(
-        &self,
-        editor: WeakView<Editor>,
-        cx: &mut ViewContext<Self>,
-    ) -> Option<Arc<Language>> {
-        let editor = editor.upgrade()?;
-        let selection = editor.read(cx).selections.newest::<usize>(cx);
-        let buffer = editor.read(cx).buffer().read(cx).snapshot(cx);
-        buffer.language_at(selection.head()).cloned()
-    }
-
     pub fn refresh_kernelspecs(&mut self, cx: &mut ModelContext<Self>) -> Task<Result<()>> {
         let kernel_specifications = kernel_specifications(self.fs.clone());
         cx.spawn(|this, mut cx| async move {
@@ -152,7 +83,7 @@ impl ReplStore {
     pub fn kernelspec(
         &self,
         language: &Language,
-        cx: &mut ViewContext<Self>,
+        cx: &mut ModelContext<Self>,
     ) -> Option<KernelSpecification> {
         let settings = JupyterSettings::get_global(cx);
         let language_name = language.code_fence_block_name();
@@ -173,101 +104,15 @@ impl ReplStore {
             .cloned()
     }
 
-    pub fn run(&mut self, editor: WeakView<Editor>, cx: &mut ViewContext<Self>) -> Result<()> {
-        if !self.enabled {
-            return Ok(());
-        }
-
-        let (selected_text, language, anchor_range) = match self.snippet(editor.clone(), cx) {
-            Some(snippet) => snippet,
-            None => return Ok(()),
-        };
-
-        let entity_id = editor.entity_id();
-
-        let kernel_specification = self
-            .kernelspec(&language, cx)
-            .with_context(|| format!("No kernel found for language: {}", language.name()))?;
-
-        let session = self.sessions.entry(entity_id).or_insert_with(|| {
-            let view =
-                cx.new_view(|cx| Session::new(editor, self.fs.clone(), kernel_specification, cx));
-            cx.notify();
-
-            let subscription = cx.subscribe(&view, |this, _session, event, _cx| match event {
-                SessionEvent::Shutdown(shutdown_event) => {
-                    this.sessions.remove(&shutdown_event.entity_id());
-                }
-            });
-
-            subscription.detach();
-
-            view
-        });
-
-        session.update(cx, |session, cx| {
-            session.execute(&selected_text, anchor_range, cx);
-        });
-
-        anyhow::Ok(())
+    pub fn get_session(&self, entity_id: EntityId) -> Option<&View<Session>> {
+        self.sessions.get(&entity_id)
     }
 
-    pub fn session(
-        &mut self,
-        editor: WeakView<Editor>,
-        cx: &mut ViewContext<Self>,
-    ) -> SessionSupport {
-        let entity_id = editor.entity_id();
-        let session = self.sessions.get(&entity_id).cloned();
-
-        match session {
-            Some(session) => SessionSupport::ActiveSession(session),
-            None => {
-                let language = self.language(editor, cx);
-                let language = match language {
-                    Some(language) => language,
-                    None => return SessionSupport::Unsupported,
-                };
-                let kernelspec = self.kernelspec(&language, cx);
-
-                match kernelspec {
-                    Some(kernelspec) => SessionSupport::Inactive(Box::new(kernelspec)),
-                    None => match language.name().as_ref() {
-                        "TypeScript" | "Python" => SessionSupport::RequiresSetup(language.name()),
-                        _ => SessionSupport::Unsupported,
-                    },
-                }
-            }
-        }
+    pub fn insert_session(&mut self, entity_id: EntityId, session: View<Session>) {
+        self.sessions.insert(entity_id, session);
     }
 
-    pub fn clear_outputs(&mut self, editor: WeakView<Editor>, cx: &mut ViewContext<Self>) {
-        let entity_id = editor.entity_id();
-        if let Some(session) = self.sessions.get_mut(&entity_id) {
-            session.update(cx, |session, cx| {
-                session.clear_outputs(cx);
-            });
-            cx.notify();
-        }
-    }
-
-    pub fn interrupt(&mut self, editor: WeakView<Editor>, cx: &mut ViewContext<Self>) {
-        let entity_id = editor.entity_id();
-        if let Some(session) = self.sessions.get_mut(&entity_id) {
-            session.update(cx, |session, cx| {
-                session.interrupt(cx);
-            });
-            cx.notify();
-        }
-    }
-
-    pub fn shutdown(&self, editor: WeakView<Editor>, cx: &mut ViewContext<Self>) {
-        let entity_id = editor.entity_id();
-        if let Some(session) = self.sessions.get(&entity_id) {
-            session.update(cx, |session, cx| {
-                session.shutdown(cx);
-            });
-            cx.notify();
-        }
+    pub fn remove_session(&mut self, entity_id: EntityId) {
+        self.sessions.remove(&entity_id);
     }
 }
