+++ crates/repl/src/runtime_panel.rs
@@ -1,20 +1,19 @@
 use crate::repl_store::ReplStore;
 use crate::{
     jupyter_settings::{JupyterDockPosition, JupyterSettings},
-    kernels::{kernel_specifications, KernelSpecification},
+    kernels::KernelSpecification,
     session::{Session, SessionEvent},
 };
 use anyhow::{Context as _, Result};
-use collections::HashMap;
 use editor::{Anchor, Editor, RangeToAnchorExt};
 use gpui::{
-    actions, prelude::*, AppContext, AsyncWindowContext, EntityId, EventEmitter, FocusHandle,
-    FocusOutEvent, FocusableView, Subscription, Task, View, WeakView, WindowContext,
+    actions, prelude::*, AppContext, AsyncWindowContext, EventEmitter, FocusHandle, FocusOutEvent,
+    FocusableView, Subscription, Task, View, WeakView,
 };
 use language::{Language, Point};
 use multi_buffer::MultiBufferRow;
 use project::Fs;
-use settings::{Settings as _, SettingsStore};
+use settings::Settings as _;
 use std::{ops::Range, sync::Arc};
 use ui::{prelude::*, ButtonLike, ElevationIndex, KeyBinding};
 use util::ResultExt as _;
@@ -29,6 +28,13 @@ actions!(
 );
 actions!(repl_panel, [ToggleFocus]);
 
+pub enum SessionSupport {
+    ActiveSession(View<Session>),
+    Inactive(Box<KernelSpecification>),
+    RequiresSetup(Arc<str>),
+    Unsupported,
+}
+
 pub fn init(cx: &mut AppContext) {
     cx.observe_new_views(
         |workspace: &mut Workspace, _cx: &mut ViewContext<Workspace>| {
@@ -36,7 +42,7 @@ pub fn init(cx: &mut AppContext) {
                 workspace.toggle_panel_focus::<RuntimePanel>(cx);
             });
 
-            workspace.register_action(|workspace, _: &RefreshKernelspecs, cx| {
+            workspace.register_action(|_workspace, _: &RefreshKernelspecs, cx| {
                 let store = ReplStore::global(cx);
                 store.update(cx, |store, cx| {
                     store.refresh_kernelspecs(cx).detach();
@@ -67,10 +73,7 @@ pub fn init(cx: &mut AppContext) {
                     let weak_editor = cx.view().downgrade();
                     panel.update(cx, |_, cx| {
                         cx.defer(|panel, cx| {
-                            let store = ReplStore::global(cx);
-                            store.update(cx, |store, cx| {
-                                store.run(weak_editor, cx).log_err();
-                            });
+                            panel.run(weak_editor, cx).log_err();
                         });
                     })
                 },
@@ -181,8 +184,11 @@ impl RuntimePanel {
                 })
             })?;
 
-            view.update(&mut cx, |this, cx| this.refresh_kernelspecs(cx))?
-                .await?;
+            view.update(&mut cx, |_panel, cx| {
+                let store = ReplStore::global(cx);
+                store.update(cx, |store, cx| store.refresh_kernelspecs(cx))
+            })?
+            .await?;
 
             Ok(view)
         })
@@ -194,6 +200,173 @@ impl RuntimePanel {
 
     fn focus_out(&mut self, _event: FocusOutEvent, cx: &mut ViewContext<Self>) {
         cx.notify();
+    }
+
+    fn snippet(
+        editor: WeakView<Editor>,
+        cx: &mut ViewContext<Self>,
+    ) -> Option<(String, Arc<Language>, Range<Anchor>)> {
+        let editor = editor.upgrade()?;
+        let editor = editor.read(cx);
+
+        let buffer = editor.buffer().read(cx).snapshot(cx);
+
+        let selection = editor.selections.newest::<usize>(cx);
+        let multi_buffer_snapshot = editor.buffer().read(cx).snapshot(cx);
+
+        let range = if selection.is_empty() {
+            let cursor = selection.head();
+
+            let cursor_row = multi_buffer_snapshot.offset_to_point(cursor).row;
+            let start_offset = multi_buffer_snapshot.point_to_offset(Point::new(cursor_row, 0));
+
+            let end_point = Point::new(
+                cursor_row,
+                multi_buffer_snapshot.line_len(MultiBufferRow(cursor_row)),
+            );
+            let end_offset = start_offset.saturating_add(end_point.column as usize);
+
+            // Create a range from the start to the end of the line
+            start_offset..end_offset
+        } else {
+            selection.range()
+        };
+
+        let anchor_range = range.to_anchors(&multi_buffer_snapshot);
+
+        let selected_text = buffer
+            .text_for_range(anchor_range.clone())
+            .collect::<String>();
+
+        let start_language = buffer.language_at(anchor_range.start)?;
+        let end_language = buffer.language_at(anchor_range.end)?;
+        if start_language != end_language {
+            return None;
+        }
+
+        Some((selected_text, start_language.clone(), anchor_range))
+    }
+
+    fn language(editor: WeakView<Editor>, cx: &mut ViewContext<Self>) -> Option<Arc<Language>> {
+        let editor = editor.upgrade()?;
+        let selection = editor.read(cx).selections.newest::<usize>(cx);
+        let buffer = editor.read(cx).buffer().read(cx).snapshot(cx);
+        buffer.language_at(selection.head()).cloned()
+    }
+
+    pub fn run(&mut self, editor: WeakView<Editor>, cx: &mut ViewContext<Self>) -> Result<()> {
+        let store = ReplStore::global(cx);
+
+        if !store.read(cx).is_enabled() {
+            return Ok(());
+        }
+
+        let (selected_text, language, anchor_range) = match Self::snippet(editor.clone(), cx) {
+            Some(snippet) => snippet,
+            None => return Ok(()),
+        };
+
+        let entity_id = editor.entity_id();
+
+        let kernel_specification = store.update(cx, |store, cx| {
+            store
+                .kernelspec(&language, cx)
+                .with_context(|| format!("No kernel found for language: {}", language.name()))
+        })?;
+
+        let session = if let Some(session) = store.read(cx).get_session(entity_id).cloned() {
+            session
+        } else {
+            let session =
+                cx.new_view(|cx| Session::new(editor, self.fs.clone(), kernel_specification, cx));
+            cx.notify();
+
+            let subscription = cx.subscribe(&session, {
+                let store = store.clone();
+                move |_this, _session, event, cx| match event {
+                    SessionEvent::Shutdown(shutdown_event) => {
+                        store.update(cx, |store, _cx| {
+                            store.remove_session(shutdown_event.entity_id());
+                        });
+                    }
+                }
+            });
+
+            subscription.detach();
+
+            store.update(cx, |store, _cx| {
+                store.insert_session(entity_id, session.clone());
+            });
+
+            session
+        };
+
+        session.update(cx, |session, cx| {
+            session.execute(&selected_text, anchor_range, cx);
+        });
+
+        anyhow::Ok(())
+    }
+
+    pub fn session(
+        &mut self,
+        editor: WeakView<Editor>,
+        cx: &mut ViewContext<Self>,
+    ) -> SessionSupport {
+        let store = ReplStore::global(cx);
+        let entity_id = editor.entity_id();
+
+        if let Some(session) = store.read(cx).get_session(entity_id).cloned() {
+            return SessionSupport::ActiveSession(session);
+        };
+
+        let language = Self::language(editor, cx);
+        let language = match language {
+            Some(language) => language,
+            None => return SessionSupport::Unsupported,
+        };
+        let kernelspec = store.update(cx, |store, cx| store.kernelspec(&language, cx));
+
+        match kernelspec {
+            Some(kernelspec) => SessionSupport::Inactive(Box::new(kernelspec)),
+            None => match language.name().as_ref() {
+                "TypeScript" | "Python" => SessionSupport::RequiresSetup(language.name()),
+                _ => SessionSupport::Unsupported,
+            },
+        }
+    }
+
+    pub fn clear_outputs(&mut self, editor: WeakView<Editor>, cx: &mut ViewContext<Self>) {
+        let store = ReplStore::global(cx);
+        let entity_id = editor.entity_id();
+        if let Some(session) = store.read(cx).get_session(entity_id).cloned() {
+            session.update(cx, |session, cx| {
+                session.clear_outputs(cx);
+            });
+            cx.notify();
+        }
+    }
+
+    pub fn interrupt(&mut self, editor: WeakView<Editor>, cx: &mut ViewContext<Self>) {
+        let store = ReplStore::global(cx);
+        let entity_id = editor.entity_id();
+        if let Some(session) = store.read(cx).get_session(entity_id).cloned() {
+            session.update(cx, |session, cx| {
+                session.interrupt(cx);
+            });
+            cx.notify();
+        }
+    }
+
+    pub fn shutdown(&self, editor: WeakView<Editor>, cx: &mut ViewContext<Self>) {
+        let store = ReplStore::global(cx);
+        let entity_id = editor.entity_id();
+        if let Some(session) = store.read(cx).get_session(entity_id).cloned() {
+            session.update(cx, |session, cx| {
+                session.shutdown(cx);
+            });
+            cx.notify();
+        }
     }
 }
 
@@ -270,7 +443,7 @@ impl Render for RuntimePanel {
     fn render(&mut self, cx: &mut ViewContext<Self>) -> impl IntoElement {
         let store = ReplStore::global(cx);
 
-        let (kernel_specifications, sessions) = store.update(cx, |store, cx| {
+        let (kernel_specifications, sessions) = store.update(cx, |store, _cx| {
             (
                 store.kernel_specifications().cloned().collect::<Vec<_>>(),
                 store.sessions().cloned().collect::<Vec<_>>(),
